tf-faster-rcnn/requirements.txt
antipathy==0.82.5
cycler==0.10.0
Cython==0.25.2
easydict==1.7
funcsigs==1.0.2
functools32==3.2.3.post2
lxml==3.8.0
matplotlib==2.0.2
mock==2.0.0
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.7
pbr==3.0.1
Pillow==4.1.1
protobuf==3.3.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
scipy==0.19.0
scription==0.80.4
six==1.10.0
subprocess32==3.2.7
tensorflow-gpu==1.1.0
Werkzeug==0.12.2
xaml==0.6.0
